+++ lib/vagrant/guest/linux.rb
@@ -57,6 +57,7 @@ module Vagrant
       def mount_folder(name, guestpath, options)
         # Determine the permission string to attach to the mount command
         options = "-o uid=`id -u #{options[:owner]}`,gid=`id -g #{options[:group]}`"
+        options += ",#{options[:extra]}" if options[:extra]
 
         attempts = 0
         while true