+++ CHANGELOG.md
@@ -41,6 +41,8 @@
     if they exist. [GH-631]
   - Added `--provision-with` to `up` which configures what provisioners run,
     by shortcut. [GH-367]
+  - Arbitrary mount options can be passed with `:extra` to any shared
+    folders. [GH-551]
   - Removed Thor as a dependency for the command line interfaces. This resulted
     in general speed increases across all command line commands.
   - Linux uses `shutdown -h` instead of `halt` to hopefully more consistently